+++ tom_build_base/lib/src/v2/core/help_generator.dart
@@ -113,6 +113,13 @@ class HelpGenerator {
       buf.writeln();
     }
 
+    // Common meta-options
+    buf.writeln('<dim>**Common Options**</dim>');
+    buf.writeln('  -h, --help                  Show this help');
+    buf.writeln('  -v, --verbose               Verbose output');
+    buf.writeln('  -n, --dry-run               Show what would be done');
+    buf.writeln();
+
     // Examples
     if (command.examples.isNotEmpty) {
       buf.writeln('<green>**Examples**</green>');